feat: create a Layout component
