+++ src/components/shared/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-cycle */
 export { default as SectionContainer } from './SectionContainer';
 export { default as Container } from './Container';
 export { default as Heading1 } from './Typography/Heading1';
@@ -8,3 +9,4 @@ export { default as Heading5 } from './Typography/Heading5';
 export { default as Body } from './Typography/Body';
 export { default as HighlightBody } from './Typography/HighlightBody';
 export { default as Button } from './Button';
+export { default as Layout } from './Layout';
